Load task-side the same namespaces loaded locally.
In order to support libraries of common task functions, we need some mechanism
to ensure that Parkour loads all necessary namespaces for a particular task,
such as those providing de/serialization multimethod & protocol implementations.
This implementation defaults to loading the same set remotely as are loaded
locally, but allows users to override as necessary for particular jobs.